+++ examples/client/gql2ts.js
@@ -6,4 +6,9 @@ gql2ts({
   root: __dirname,
   introspection: { url: "http://localhost:4000/graphql", fetch },
   client: true,
+  scalarTypes: {
+    UUID: "string",
+    JSON: "any",
+    ESNumber: "number",
+  },
 });